requirements.txt
pymediainfo==6.1.0
PyYAML==6.0.2
